#5 Added token generator Reworked FileController download actions Simplified id acquisition in AccountController and FileController

+++ UploadIt.Services/Security/ITokenGenerator.cs
@@ -10,6 +10,6 @@ namespace UploadIt.Services.Security
 {
     public interface ITokenGenerator
     {
-        TokenInfo GenerateJwtToken(string secret, Claim[] claims, int minutesExpirationTime);
+        TokenInfo GenerateJwtToken(string secret, Claim[] claims, double minutesExpirationTime);
     }
 }
